feat: incorporate PriorityQueue from .NET 6 (Preview)
Implement PriorityQueue based on titles and number of views for priority management,
a classic SortedList has already been tested with but the need for custom
IComparer to allow duplicate keys makes it a broken solution. Also we need a better
solution to save the queue in descending order, the combination of IComparer
and TryPeek is not good.
Wikipedia dumps server does not allow more than 2 (i think so) concurrent connections,
we cannot implement an "efficient" producer/consumer solution with more than 2
parallel downloads.
Split the solution in files with tests.