feat: add DI to host (console app) with addlogging and addhttpclient
Add Serilog and mute Microsoft* and System* to log event level of warning.
Add httpclientfactory although in this case is unnecesary because
the solution has a short lifetime duration, so is very difficult
to provoke port exhaustion issues.
Add DI with hostbuilder also it can be implemented with IServiceCollection.

Looking forward to persist the in-memory cache (dictionary) with sqlite¿?
and implement a very basic monitoring.