feat: add documentation and fix sqlite implementation
Add documentation without the monitoring part because is not implemented
yet.
Add host (console project) to the Dockerfile, we are not using it in the CI at the
moment but we could use the publish part as an artifact in the pipeline.
Fix sqlite implementation adding an absolute path to SqliteConnectionStringBuilder.

Looking forward to implement some sort of observability around the solution.